Resolutions:
- CHANGELOG.md: dropped the FIDO MDS entry entirely, per review. This is
  groundwork with no user-visible effect, and the release notes will cover
  the feature once enforcement lands.
- src/bin/Cargo.toml: kept upstream's new 'profiling' feature next to our
  second [[bin]] target, and removed the 'default-run' comment.
- migrations: 32_ and V27__ are both claimed upstream now
  (32_one_time_password, V27__one_time_password), so the FIDO MDS schema
  moves to 35_fido_mds.sql and V30__fido_mds.sql. Neither of ours has ever
  shipped in a release, so renumbering is safe.

+++ src/api_types/src/generic.rs
@@ -110,8 +110,8 @@ pub struct SearchParams {
     pub ty: SearchParamsType,
     /// Index
     pub idx: SearchParamsIdx,
-    /// The actual search query - validation: `[a-zA-Z0-9,.:/_\-&?=~#!$'()*+%@]+`
-    #[validate(regex(path = "*RE_SEARCH", code = "[a-zA-Z0-9,.:/_\\-&?=~#!$'()*+%@]+"))]
+    /// The actual search query - validation: `[a-zA-Z0-9,.:/_\-&?=~#!$'()*+@]+`
+    #[validate(regex(path = "*RE_SEARCH", code = "[a-zA-Z0-9,.:/_\\-&?=~#!$'()*+@]+"))]
     pub q: String,
     pub limit: Option<u16>,
 }
@@ -173,7 +173,7 @@ pub struct I18nConfigResponse {
 #[derive(Serialize, ToSchema)]
 pub struct LoginTimeResponse {
     pub argon2_params: Argon2ParamsResponse,
-    pub login_time: u32,
+    pub login_time: i64,
     pub num_cpus: usize,
 }
